+++ src/file.rs
@@ -3,27 +3,27 @@ use std::fs::File;
 
 pub struct FileReader<'a> {
     file: &'a File,
-    file_size: u64,
-    chunk_size: u64,
+    file_size: usize,
+    chunk_size: usize,
 }
 
 impl<'a> FileReader<'a> {
-    pub fn new(file: &File, chunk_size: u64) -> FileReader {
+    pub fn new(file: &File, chunk_size: usize) -> FileReader {
         FileReader {
             file,
-            file_size: file.metadata().unwrap().len(),
+            file_size: file.metadata().unwrap().len() as usize,
             chunk_size,
         }
     }
 
-    pub fn get_chunk(&self, page: u64) -> Option<Chunk> {
+    pub fn get_chunk(&self, page: usize) -> Option<Chunk> {
         let offset = page * self.chunk_size;
         if offset >= self.file_size {
             return None;
         }
         let mut mmap_option = MmapOptions::new();
         mmap_option
-            .offset(offset)
+            .offset(offset as u64)
             .len(std::cmp::min(self.chunk_size, self.file_size - offset) as usize);
 
         Some(Chunk {
@@ -32,19 +32,19 @@ impl<'a> FileReader<'a> {
         })
     }
 
-    pub fn is_page_available(&self, page: u64) -> bool {
+    pub fn is_page_available(&self, page: usize) -> bool {
         page * self.chunk_size >= self.file_size
     }
 }
 
 pub struct FileWriter<'a> {
     file: &'a File,
-    file_size: u64,
-    chunk_size: u64,
+    file_size: usize,
+    chunk_size: usize,
 }
 
 impl<'a> FileWriter<'a> {
-    pub fn new(file: &File, file_size: u64, chunk_size: u64) -> FileWriter {
+    pub fn new(file: &File, file_size: usize, chunk_size: usize) -> FileWriter {
         file.set_len(file_size as u64).unwrap();
         FileWriter {
             file,
@@ -53,7 +53,7 @@ impl<'a> FileWriter<'a> {
         }
     }
 
-    pub fn get_chunk_mut(&self, page: u64) -> Option<ChunkMut> {
+    pub fn get_chunk_mut(&self, page: usize) -> Option<ChunkMut> {
         let offset = page * self.chunk_size;
         if offset >= self.file_size {
             return None;
@@ -69,18 +69,18 @@ impl<'a> FileWriter<'a> {
         })
     }
 
-    pub fn is_page_available(&self, page: u64) -> bool {
+    pub fn is_page_available(&self, page: usize) -> bool {
         page * self.chunk_size >= self.file_size
     }
 }
 
 pub struct Chunk {
-    pub page: u64,
+    pub page: usize,
     pub mmap: Mmap,
 }
 pub struct ChunkMut {
-    pub page: u64,
+    pub page: usize,
     pub mmap_mut: MmapMut,
 }
 
-pub const TAG_LEN: u64 = 16;
+pub const TAG_LEN: usize = 16;
